requirements.txt
colored==1.4.3
docopt==0.6.2
mc-bin-client==1.0.1
PyYAML==6.0
